fix(PR Review): refactor commonification
