+++ erpnext/accounts/doctype/account_group/account_group_dashboard.py
@@ -1,0 +1,13 @@
+from frappe import _
+
+
+def get_data():
+	return {
+		'fieldname': 'account_group',
+		'transactions': [
+			{
+				'label': _('Used In'),
+				'items': ['Account Group']
+			}
+		]
+	}
